+++ apps/petrinaut-website/package.json
@@ -21,11 +21,11 @@
     "@hashintel/ds-helpers": "workspace:*",
     "@hashintel/petrinaut": "workspace:*",
     "@hashintel/petrinaut-core": "workspace:*",
+    "@local/petrinaut-optimizer-client": "workspace:*",
     "@mantine/hooks": "8.3.5",
     "@pandacss/dev": "1.11.1",
     "@sentry/react": "10.64.0",
     "ai": "6.0.182",
-    "eventsource-parser": "3.0.8",
     "immer": "10.1.3",
     "react": "19.2.6",
     "react-dom": "19.2.6",
