+++ libs/@local/petrinaut-optimizer-client/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@local/petrinaut-optimizer-types",
+  "name": "@local/petrinaut-optimizer-client",
   "version": "0.0.0-private",
   "private": true,
-  "description": "Generated TypeScript types for the Petrinaut optimizer API",
+  "description": "Isomorphic client utilities and generated types for Petrinaut Optimizer",
   "license": "AGPL-3.0",
   "type": "module",
   "main": "dist/index.js",
@@ -17,7 +17,11 @@
     "build": "rimraf dist && tsc --build tsconfig.build.json",
     "codegen": "openapi-typescript ../../../apps/petrinaut-opt/openapi/openapi.json --default-non-nullable false --output src/openapi.gen.ts && oxfmt --write --ignore-path /dev/null src/openapi.gen.ts",
     "lint:tsc": "tsc --noEmit",
-    "test:unit": "yarn codegen && git diff --exit-code -- src/openapi.gen.ts"
+    "test:unit": "yarn codegen && git diff --exit-code -- src/openapi.gen.ts && vitest --run src"
+  },
+  "dependencies": {
+    "@hashintel/petrinaut-core": "workspace:*",
+    "eventsource-parser": "3.0.8"
   },
   "devDependencies": {
     "@apps/petrinaut-opt": "workspace:*",
@@ -25,6 +29,7 @@
     "openapi-typescript": "7.13.0",
     "oxfmt": "0.50.0",
     "rimraf": "6.1.3",
-    "typescript": "5.9.3"
+    "typescript": "5.9.3",
+    "vitest": "4.1.8"
   }
 }
