+++ libs/@local/petrinaut-optimizer-client/src/index.ts
@@ -1,0 +1,3 @@
+export { decodePetrinautOptimizerStream } from "./decode-optimization-stream.js";
+export type { DecodePetrinautOptimizerStreamOptions } from "./decode-optimization-stream.js";
+export type { components, operations, paths, webhooks } from "./openapi.gen.js";
