+++ libs/@local/petrinaut-optimizer-client/src/openapi.gen.ts
