+++ libs/@local/petrinaut-optimizer-client/tsconfig.build.json
@@ -1,5 +1,6 @@
 {
   "extends": "./tsconfig.json",
+  "exclude": ["src/**/*.test.ts"],
   "compilerOptions": {
     "declaration": true,
     "outDir": "dist",
