+++ libs/@local/petrinaut-optimizer-client/tsconfig.json
