--- libs/@local/petrinaut-optimizer-types/src/index.ts
@@ -1,1 +1,0 @@
-export type { components, operations, paths, webhooks } from "./openapi.gen.js";
